Access InputDeviceIdentifier via reflection when needed
To avoid hitting SDK compatibility checks on Android 9, access
InputDeviceIdentifier via a reflection.

This makes us hit the check only when getIdentifier() is actually
called, which should not happen on Android 9.

InputDeviceIdentifier is needed for some pre-Android-9 BlackBerry
phones that have no other way to access the layout selection screen.