Add empty layout as FinQwerty None
This can be used to get the stock layout available in the Android quick
layout switcher in case one is not provided by the system.

IOW, selecting this layout is the same as not selecting any layouts (in
which case layouts cannot be switched using the switcher).

Suggested by Eske Rahn.