pro1: Added Polish keyboard layout (#9)
The Polish programmer's keyboard layout is the de-facto standard layout
in Poland and is used with regular US (International) QWERTY keyboards.

The 6 extra characters in the Polish alphabet are accessible by using
AltGr on the respective similarily looking standard characters (except
for ź being on x, as ż is already on z). I left the rest of the
original US International layout intact as one may want to use this
layout for multiple (european) languages.