pro1: Add missing non-dead ^ and ~ to various QWERTZ layouts
Some of the physical QWERTZ layouts for Pro1 are missing non-dead '^'
and '~'. This is a problem as inputting e.g. dead ~ twice or dead ~ with
space actually does produce a different character than non-dead ~ at
least on stock Android on Pro1.

Add non-dead ^ to alt+W (where free) and non-dead ~ to alt+Ü (where
free), which also matches key labels.

Specifically:
- alt+W = ^: cze, dan, fin, ger, swe, usa
- alt+Ü = ~: cze, dan, fin, ger, swe